+++ src/main/java/fr/insa/empire/treillis/Barre.java
@@ -1,10 +1,11 @@
 package fr.insa.empire.treillis;
 
+//import fr.insa.empire.utils.Identificateur;
+
 import fr.insa.empire.utils.Identificateur;
 
 import java.io.BufferedWriter;
 import java.io.IOException;
-import java.util.Set;
 
 public class Barre {
