+++ src/main/java/fr/insa/empire/treillis/Noeud_appui.java
@@ -1,7 +1,8 @@
 package fr.insa.empire.treillis;
 
+import fr.insa.empire.utils.Identificateur;
+
 import java.io.BufferedWriter;
-import java.io.IOException;
 
 public class Noeud_appui extends Noeuds {
